+++ db2/history2/operation_details.go
@@ -54,6 +54,11 @@ type OperationDetails struct {
 	RemoveAssetPair            *RemoveAssetPairDetails            `json:"remove_asset_pair,omitempty"`
 	InitiateKYCRecovery        *InitiateKYCRecoveryDetails        `json:"initiate_kyc_recovery,omitempty"`
 	CreateKYCRecoveryRequest   *CreateKYCRecoveryRequestDetails   `json:"create_kyc_recovery_request,omitempty"`
+	CreateManageOfferRequest   *CreateManageOfferRequestDetails   `json:"create_manage_offer_request,omitempty"`
+	CreatePaymentRequest       *CreatePaymentRequestDetails       `json:"create_payment_request,omitempty"`
+	RemoveAsset                *RemoveAssetDetails                `json:"remove_asset,omitempty"`
+	OpenSwap                   *OpenSwapDetails                   `json:"open_swap,omitempty"`
+	CloseSwap                  *CloseSwapDetails                  `json:"close_swap,omitempty"`
 }
 
 //Value - converts operation details into jsonb
@@ -528,4 +533,51 @@ type CreateKYCRecoveryRequestDetails struct {
 	CreatorDetails regources.Details            `json:"creator_details"`
 	AllTasks       *uint32                      `json:"all_tasks"`
 	RequestDetails RequestDetails               `json:"request_details"`
+}
+
+type CreateManageOfferRequestDetails struct {
+	ManageOfferDetails ManageOfferRequest `json:"manage_offer_details"`
+	AllTasks           *uint32            `json:"all_tasks"`
+	RequestDetails     RequestDetails     `json:"request_details"`
+}
+
+type PaymentRequestDetails struct {
+	AccountFrom             string           `json:"account_from"`
+	BalanceFrom             string           `json:"balance_from"`
+	Amount                  regources.Amount `json:"amount"`
+	SourceFee               regources.Fee    `json:"source_fee"`
+	DestinationFee          regources.Fee    `json:"destination_fee"`
+	SourcePayForDestination bool             `json:"source_pay_for_destination"`
+	Subject                 string           `json:"subject"`
+	Reference               string           `json:"reference"`
+}
+
+type CreatePaymentRequestDetails struct {
+	PaymentDetails PaymentRequestDetails `json:"payment_details"`
+	AllTasks       *uint32               `json:"all_tasks"`
+	RequestDetails RequestDetails        `json:"request_details"`
+}
+
+type RemoveAssetDetails struct {
+	Code string `json:"code"`
+}
+
+type OpenSwapDetails struct {
+	AccountFrom             string            `json:"account_from"`
+	AccountTo               string            `json:"account_to"`
+	BalanceFrom             string            `json:"balance_from"`
+	BalanceTo               string            `json:"balance_to"`
+	Amount                  regources.Amount  `json:"amount"`
+	Asset                   string            `json:"asset"`
+	SourceFee               regources.Fee     `json:"source_fee"`
+	DestinationFee          regources.Fee     `json:"destination_fee"`
+	SourcePayForDestination bool              `json:"source_pay_for_destination"`
+	SecretHash              string            `json:"secret_hash"`
+	LockTime                time.Time         `json:"lock_time"`
+	Details                 regources.Details `db:"details"`
+}
+
+type CloseSwapDetails struct {
+	ID     int64   `json:"id"`
+	Secret *string `json:"secret"`
 }