+++ CHANGELOG.md
@@ -1,4 +1,62 @@
+# Changelog
+
 ## Unreleased
+
+## 3.6.2
+
+### Added
+
+* Cache for `GET` requests
+
+### Fixed
+
+* Batch inserter corner case
+
+## 3.6.1
+
+### Added
+
+* `/v3/license` endpoint to return current license info
+
+### Removed
+
+* Request for account endpoint (`/v3/accounts/{id}/requests/{request_id}`)
+
+### Fixed
+
+* RequestID in create issuance request operation details
+* includes for swaps (`/v3/swaps`)
+* docs consistency with actual responses
+
+## 3.6.0
+
+### Added
+
+* `/v3/manage_offer_requests` endpoint
+* `/v3/create_payment_requests` endpoint
+* `CreateManageOfferRequestOp` ingestion
+* `CreatePaymentRequestOp` ingestion
+* `CreatePaymentRequestOp` ingestion
+* `PaymentRequest` ingestion
+* `ManageOfferRequest` ingestion
+* `RemoveAssetOp` ingestion
+* `OpenSwapOp` ingestion
+* `CloseSwapOp` ingestion
+* `Swap` ingestion
+* `/v3/swaps` endpoint to get filtered list of swaps
+* `/v3/swaps/{id}` endpoint to get swap by id
+
+### Fixed
+
+* panic on `v3/balances`
+* error on reviewable request ingestion
+
+## 3.5.2
+
+### Fixed
+
+* Key for atomic swap ask quote asset relation
+* Panic on `ManagePoll` ingestion
 
 # 3.5.1.2